feat(ew): ability to get progress via iframe
